# docs/swagger.yaml
basePath: /v1
definitions:
...
  github_com_pickle_pw_monolith_internal_types.ImagePreviewRes:
    properties:
      previewId:
        type: string
      previewUrl:
        type: string
    type: object
  github_com_pickle_pw_monolith_internal_types.ImageRes:
    properties:
      url:
        type: string
    type: object
...
  github_com_pickle_pw_monolith_internal_types.PosterPreviewRes:
    properties:
      createdAt:
        type: string
      id:
        type: string
      url:
        type: string
    type: object
...
host: api.pickle.pw
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Pickle server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pickle API
  version: "1.0"
paths:
...
  /v1/profiles/{username}:
    get:
      consumes:
      - application/json
      description: Get profile by username
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Avatar size
        in: query
        name: avatarSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get profile by username
      tags:
      - profiles
...
  /v1/users/{userId}/posters/previews:
    get:
      consumes:
      - application/json
      description: Get all poster previews
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - default: lg
        description: Size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.PosterPreviewRes'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all poster previews
      tags:
      - posters
    post:
      consumes:
      - application/json
      description: Upload a poster preview
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: URL
        in: formData
        name: url
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.ImagePreviewRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Upload a poster preview
      tags:
      - posters
  /v1/users/{userId}/posters/previews/{previewId}:
    delete:
      consumes:
      - application/json
      description: Delete a poster preview
      parameters:
      - description: Preview ID
        in: path
        name: previewId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a poster preview
      tags:
      - posters
    get:
      consumes:
      - application/json
      description: Get a poster preview
      parameters:
      - description: Preview ID
        in: path
        name: previewId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.PosterPreviewRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a poster preview
      tags:
      - posters
  /v1/users/{userId}/posters/previews/{previewId}/image:
    get:
      consumes:
      - application/json
      description: Get a poster preview image URL
      parameters:
      - description: Preview ID
        in: path
        name: previewId
        required: true
        type: string
      - description: Size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.ImageRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a poster preview image URL
      tags:
      - posters
...
swagger: "2.0"